+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Top proposals calculation
+
 ## [0.0.39] - 2023-10-04
 
 ### Fixed